requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
docutils==0.19
idna==3.3
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==0.104.2
Pillow==9.2.0
Pygments==2.12.0
pyserial==3.5
requests==2.28.1
urllib3==1.26.9
